# perf: remove repeated calls to expensive function (and other improvements)