In the Sales Invoice List View:

- `prepare_args` used to take more than 90% of time, actual query taking only around 9%

    ![image](https://github.com/user-attachments/assets/2de5d976-ef8c-4475-abac-103eab457e2a)

- of this, `apply_fieldlevel_read_permissions` used to take more than 90%

    ![image](https://github.com/user-attachments/assets/8edaafbb-e99a-4464-95c2-757cf8bd06f2)

- This was the payload:
    
    ```py
    {'doctype': 'Sales Invoice', 'fields': '["`tabSales Invoice`.`name`","`tabSales Invoice`.`owner`","`tabSales Invoice`.`creation`","`tabSales Invoice`.`modified`","`tabSales Invoice`.`modified_by`","`tabSales Invoice`.`_user_tags`","`tabSales Invoice`.`_comments`","`tabSales Invoice`.`_assign`","`tabSales Invoice`.`_liked_by`","`tabSales Invoice`.`docstatus`","`tabSales Invoice`.`idx`","`tabSales Invoice`.`total`","`tabSales Invoice`.`net_total`","`tabSales Invoice`.`total_taxes_and_charges`","`tabSales Invoice`.`grand_total`","`tabSales Invoice`.`rounding_adjustment`","`tabSales Invoice`.`rounded_total`","`tabSales Invoice`.`total_advance`","`tabSales Invoice`.`outstanding_amount`","`tabSales Invoice`.`discount_amount`","`tabSales Invoice`.`total_billing_amount`","`tabSales Invoice`.`paid_amount`","`tabSales Invoice`.`change_amount`","`tabSales Invoice`.`write_off_amount`","`tabSales Invoice`.`status`","`tabSales Invoice`.`customer_name`","`tabSales Invoice`.`customer`","`tabSales Invoice`.`base_grand_total`","`tabSales Invoice`.`due_date`","`tabSales Invoice`.`company`","`tabSales Invoice`.`currency`","`tabSales Invoice`.`is_return`","`tabSales Invoice`.`_seen`","`tabSales Invoice`.`party_account_currency`"]', 'filters': '[]', 'order_by': '`tabSales Invoice`.creation desc', 'start': '0', 'page_length': '20', 'view': 'List', 'group_by': '', 'with_comment_count': '1', 'cmd': 'frappe.desk.reportview.get'}
    ```
    
    It had 35 fields from the `Sales Invoice` doctype, all containing "dot".
- all fields containing dot were assumed to be child fields, with `get_permitted_fields` being called for each 😄 
- this led to O(n) + 1 (base) calls to an expensive function

This PR fixes that logic and makes some common sense constants and sets for quicker lookup.

#### Before

```py
In [39]: %timeit -n300 get()
3.88 ms ± 82.3 µs per loop (mean ± std. dev. of 7 runs, 300 loops each)
```

#### After

```py
In [45]: %timeit -n300 get()
867 µs ± 27.7 µs per loop (mean ± std. dev. of 7 runs, 300 loops each)
```